requirements.txt
streamlit==1.32.0

python-dotenv==1.0.1
pandas==2.1.4
requests==2.31.0
beautifulsoup4==4.12.3
lxml==4.9.4

pdfplumber==0.10.3
PyPDF2==3.0.1
reportlab==4.1.0
plotly==5.18.0

chromadb==0.4.24
langchain==0.1.12

groq==0.9.0
langchain-groq==0.1.5
httpx==0.27.0

sentence-transformers==2.6.1   # for embeddings
torch>=2.0                     # required for sentence-transformers
